Revert "feat: move __tests__ and __mocks__ under test/unit folder"
This reverts commit 3c1eb3d4cb4b8164a6d9f1d5139df5ac40aef395.
